+++ RELEASES.md
@@ -72,6 +72,7 @@ Upcoming
 - Prevent Show Tibco Queues to still fail when passive servers in server url
 - Fix servlets in web.xml causing errors when cmis dependency is missing
 - Bugfix error "ShowConfigurationStatus.xsl line [683]: Cannot convert string to double"
+- Recover job: restart RestListeners of which uriPattern is not registered correctly